+++ app/src/main/java/com/ysw/weatheralarm/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.ysw.weatherAlram
+package com.ysw.weatheralarm
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.ysw.weatherAlram.ui.theme.MyApplicationTheme
+import com.ysw.weatheralarm.ui.theme.MyApplicationTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
