+++ app/src/main/java/com/ysw/weatheralarm/ui/theme/Theme.kt
@@ -1,6 +1,5 @@
-package com.ysw.weatherAlram.ui.theme
+package com.ysw.weatheralarm.ui.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
