+++ app/src/main/java/com/ysw/weatheralarm/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.ysw.weatherAlram.ui.theme
+package com.ysw.weatheralarm.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
